Update webpages style and layout
Group the contents of the webpage by putting it inside a div container. Add margins, padding, width, and height to the div containers. Set also the text-align property of the content to justify to organize contents. Add the hover and active selectors to the anchor elements in the home page.